Firebase Scripts
I have connected the database to our web app, and started on user sign in and user sign up scripts.

This page will essentially be full of scripts that is catered for databse interaction, so for functions that require database interaction such as sign up, get profile etc, please remember to call the script with type = "module" at a location before any call of functions that uses firebase. thanks